Enable jshint and fix style errors.
jshint was configured to run by default on 'grunt' without any arguments,
but 'npm test' only runs the unit tests without linting. This change
runs jshint when you run 'npm test'.

This uncovered several lint errors, which are also fixed in this change.

Finally, I noticed that running tests produces a new file
test/files/output.xml that should not be checked in, so I added it as an
exclusion in .gitignore.